+++ packages/components/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @scaffold-ui/components
 
+## 0.1.2-rc.5
+
+### Patch Changes
+
+- 8a1385f: fix refocus scroll on Input components
+  update collapsible component to use rounded-2xl
+  have default font family for all components
+  fix base input component colors
+
+## 0.1.2-rc.4
+
+### Patch Changes
+
+- 5e5778b: make address optional for useBalance and Balance component
+- 5e5778b: ui polish to debug component
+- 5e5778b: Address component blockexplorer link and "use client" for all components
+- Updated dependencies [5e5778b]
+  - @scaffold-ui/hooks@0.1.2-rc.1
+
+## 0.1.2-rc.3
+
+### Patch Changes
+
+- bd0d241: UI polish
+
+## 0.1.2-rc.2
+
+### Patch Changes
+
+- 211c108: make Address size prop optional
+
+## 0.1.2-rc.1
+
+### Patch Changes
+
+- b3da6fe: setup themeing for components package
+
+## 0.1.2-rc.0
+
+### Patch Changes
+
+- 4baa885: feat: have common BaseInput for Input components
+- Updated dependencies [6674b6e]
+  - @scaffold-ui/hooks@0.1.2-rc.0
+
 ## 0.1.1
 
 ### Patch Changes